catch an odd divide-by-zero bug
not exactly sure why it happens, but occasionally when struck with some amount of ion damage at 0 energy (or with no energy capacity), "shieldMaintenance" gets set to "-4inf" if I caught it correctly, which then sets shields, energy, fuel, and hull to "nan" and effectively kills you. Because "-4inf" is a number, and is negative, this will set it to 0 and fix the issue.